2. Frameworks e Bibliotecas Populares/Web/FastAPI/requirements.txt
fastapi[standard]
pydantic
pyjwt
pwdlib[argon2]
python-multipart